+++ library/src/main/java/ru/yoomoney/sdk/kassa/payments/ui/ConfirmationActivity.kt
@@ -29,7 +29,6 @@ import com.yandex.metrica.YandexMetrica
 import ru.yoomoney.sdk.kassa.payments.BuildConfig
 import ru.yoomoney.sdk.kassa.payments.checkoutParameters.PaymentMethodType
 import ru.yoomoney.sdk.kassa.payments.checkoutParameters.TestParameters
-import ru.yoomoney.sdk.kassa.payments.di.CheckoutInjector
 import ru.yoomoney.sdk.kassa.payments.logging.ReporterLogger
 import ru.yoomoney.sdk.kassa.payments.metrics.Reporter
 import ru.yoomoney.sdk.kassa.payments.metrics.SberPayConfirmationStatusSuccess
@@ -38,7 +37,6 @@ import ru.yoomoney.sdk.kassa.payments.utils.INVOICING_AUTHORITY
 import ru.yoomoney.sdk.kassa.payments.utils.SBERPAY_PATH
 import ru.yoomoney.sdk.kassa.payments.utils.createSberbankIntent
 import ru.yoomoney.sdk.kassa.payments.utils.getSberbankPackage
-import javax.inject.Inject
 
 internal const val EXTRA_PAYMENT_METHOD_TYPE = "ru.yoomoney.sdk.kassa.payments.extra.PAYMENT_METHOD_TYPE"
 internal const val EXTRA_CONFIRMATION_URL = "ru.yoomoney.sdk.kassa.payments.extra.EXTRA_SBER_CONFIRMATION_URL"
@@ -46,9 +44,6 @@ internal const val EXTRA_CONFIRMATION_URL = "ru.yoomoney.sdk.kassa.payments.extr
 private const val SBER_PAY_CONFIRMATION_ACTION = "actionSberPayConfirmation"
 
 internal class ConfirmationActivity : AppCompatActivity() {
-
-    @Inject
-    lateinit var testParameters: TestParameters
 
     private var isWaitingForResult = false
 
@@ -60,12 +55,12 @@ internal class ConfirmationActivity : AppCompatActivity() {
     private lateinit var reporter: Reporter
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        CheckoutInjector.inject(this)
+        val testParameters = intent.getParcelableExtra(EXTRA_TEST_PARAMETERS) as TestParameters?
         reporter = ReporterLogger(
             YandexMetricaReporter(
                 YandexMetrica.getReporter(applicationContext, BuildConfig.APP_METRICA_KEY)
             ),
-            testParameters.showLogs,
+            testParameters?.showLogs ?: false,
             this
         )
         val confirmationUrl = intent.getStringExtra(EXTRA_CONFIRMATION_URL)
@@ -74,7 +69,7 @@ internal class ConfirmationActivity : AppCompatActivity() {
             startConfirmationProcess(
                 confirmationUrl,
                 intent.getSerializableExtra(EXTRA_PAYMENT_METHOD_TYPE) as PaymentMethodType,
-                requireNotNull(intent.getParcelableExtra(EXTRA_TEST_PARAMETERS))
+                testParameters
             )
             return
         } else {
@@ -94,9 +89,21 @@ internal class ConfirmationActivity : AppCompatActivity() {
         }
     }
 
-    private fun startConfirmationProcess(confirmationUrl: String, paymentMethodType: PaymentMethodType, testParameters: TestParameters) {
+    private fun startConfirmationProcess(
+        confirmationUrl: String,
+        paymentMethodType: PaymentMethodType,
+        testParameters: TestParameters?,
+    ) {
         when (paymentMethodType) {
-            PaymentMethodType.SBERBANK -> startActivity(createSberbankIntent(this, confirmationUrl, getSberbankPackage(testParameters.hostParameters.isDevHost)))
+            PaymentMethodType.SBERBANK -> {
+                startActivity(
+                    createSberbankIntent(
+                        this,
+                        confirmationUrl,
+                        getSberbankPackage(testParameters?.hostParameters?.isDevHost ?: false)
+                    )
+                )
+            }
             else -> {
                 setResult(Activity.RESULT_CANCELED)
                 finish()